+++ src/common/assets.ts
@@ -1,7 +1,0 @@
-export class APP_IMAGE {
-    static notebook = require('../../assets/notebook.png')
-}
-
-export class APP_ICONS {
-    static plus = require('../../assets/plus.png')
-}