+++ src/common/icons.ts
@@ -1,0 +1,3 @@
+export default class APP_ICONS {
+    static plus = require('../../assets/plus.png')
+}